@@ -77,13 +77,13 @@
     "@types/proper-lockfile": "^4.1.4",
     "@types/semver": "^7.5.8",
     "@types/splitpanes": "^2.2.1",
-    "@typescript-eslint/eslint-plugin": "^7.18.0",
-    "@typescript-eslint/parser": "^7.18.0",
+    "@typescript-eslint/eslint-plugin": "^8.44.0",
+    "@typescript-eslint/parser": "^8.44.0",
     "@vitejs/plugin-vue": "^6.0.0",
     "@vitest/coverage-v8": "^3.1.1",
     "@vitest/ui": "^3.1.1",
-    "@vue/eslint-config-prettier": "^9.0.0",
-    "@vue/eslint-config-typescript": "^13.0.0",
+    "@vue/eslint-config-prettier": "^10.2.0",
+    "@vue/eslint-config-typescript": "^14.6.0",
     "@vue/test-utils": "^2.2.1",
     "archiver": "^7.0.1",
     "concurrently": "^9.0.1",
@@ -91,11 +91,11 @@
     "electron": "^38.1.0",
     "electron-builder": "^26.0.12",
     "electron-devtools-installer": "^4.0.0",
-    "eslint": "^8.26.0",
+    "eslint": "^9.36.0",
     "eslint-config-prettier": "^10.0.1",
-    "eslint-import-resolver-typescript": "^3.5.2",
+    "eslint-import-resolver-typescript": "^4.4.4",
     "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-vue": "^9.7.0",
+    "eslint-plugin-vue": "^9.33.0",
     "jsdom": "^26.0.0",
     "license-checker": "^25.0.1",
     "prettier": "^3.0.1",
@@ -104,6 +104,7 @@
     "tsc-alias": "^1.7.0",
     "tsx": "^4.19.3",
     "typescript": "^5.5.4",
+    "typescript-eslint": "^8.44.1",
     "vite": "^7.1.5",
     "vite-plugin-pwa": "^1.0.0",
     "vitest": "^3.1.1",
