@@ -5,7 +5,7 @@ export function fileURLToPath(fileURL: string, defaultPath: string): string {
   if (fileURL) {
     try {
       return url.fileURLToPath(fileURL);
-    } catch (e) {
+    } catch {
       getAppLogger().warn(`invalid file URL: ${fileURL}`);
     }
   }
