@@ -59,7 +59,7 @@ export function setup(): void {
           });
         });
       }
-    } catch (e) {
+    } catch {
       return;
     }
     try {
